u8815: add script for determining libcm to be used
This script thanks to a modification to fs/proc/app_info.c links
the correct libcm.so to /system/lib/libcm.so based on which baseband
is being used. This is possible as the baseband can be identified
by what framebuffer memory has been allocated by the bootloader.

This is being done because many people have been suffering from
random reboots on any custom ROM on the 2030 baseband for some
unknown reason, and 109808 works fine. Asking everyone to move
back to the 109808 baseband would be unreasonable and having
to maintain 2 builds for 1 lib is unreasonable too. :P

Thanks kra1o5 for the libcm.so findings :D