setbtmac: get the bdaddr from the radio instead of cheating
Cheating worked unless you were on the Gingerbread bootloader, as
the serialno was not set to the bdaddr but some other weird string.

This is pretty much the same method that brcm_patchram_plus uses, without
all of the fancy structs and enums... and error checking, but it's hard to do
that without having the code for this stuff, and probably not really necessary
anyway. :D

Also clean up setwifimac