Make internal SD (emmc) the primary storage
Necessary to fix the CMUpdater/recovery disaster post 4.3 without hacking
away at recovery code as everything assumes this kind of setup.
It's also "mandatory for CM10", so maybe we're a little late here. :D

When (if) the vold switchable pair stuff gets merged this shouldn't be much of
a problem, but I'm sure there are going to be "vocal" users following this change. :(